test: separate Valgrind runner from FFI workload
